fixed the type error
